simplified bond protection structure
